fix metadata issue
